stepik-boomstream/requirements.txt
flask
sqlalchemy
psycopg2-binary
requests
gunicorn
werkzeug
python-dotenv
